+++ stick/stick.hpp
@@ -7,10 +7,10 @@ public:
 
     struct stick_vertex{
         Vector2 point1;  
-        Vector2 point2;  
+        Vector2 point2; 
     };
     stick_vertex st;
 
-    void Update(Vector2& point1, Vector2& point2, int length);
+    void Update(Vector2& point1, Vector2& point2, float length);
     void Render(Vector2& point1, Vector2& point2);
 };
